Move `getSortedEigenpairs`, `extractEigenvaluesFromEigenpairs`, `extractEigenvectorsFromEigenpairs` to `LinearAlgebraUtils.java`
